LPS-58680 match application-list-api
