+++ modules/apps/configuration-admin/configuration-admin-equinox/src/com/liferay/configuration/admin/equinox/EquinoxExtendedObjectClassDefinition.java
@@ -14,8 +14,8 @@
 
 package com.liferay.configuration.admin.equinox;
 
-import com.liferay.configuration.admin.api.ExtendedAttributeDefinition;
-import com.liferay.configuration.admin.api.ExtendedObjectClassDefinition;
+import com.liferay.configuration.admin.ExtendedAttributeDefinition;
+import com.liferay.configuration.admin.ExtendedObjectClassDefinition;
 
 import java.io.IOException;
 import java.io.InputStream;
